requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask_jwt_extended
requests
flask-migrate
psycopg2